pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "moc-plus"
version = "0.1.0"
description = "A command-line music player"
readme = "README.md"
requires-python = ">=3.8"
dependencies = [
    "beautifulsoup4==4.13.4",
    "certifi==2025.8.3",
    "charset-normalizer==3.4.3",
    "idna==3.10",
    "linkify-it-py==2.0.3",
    "lxml==6.0.0",
    "markdown-it-py==3.0.0",
    "mdit-py-plugins==0.4.2",
    "mdurl==0.1.2",
    "platformdirs==4.3.8",
    "Pygments==2.19.2",
    "python-mpv==1.0.8",
    "requests==2.32.4",
    "rich==14.1.0",
    "soupsieve==2.7",
    "textual==5.3.0",
    "typing_extensions==4.14.1",
    "uc-micro-py==1.0.3",
    "urllib3==2.5.0",
]

[project.scripts]
moc-plus = "moc_plus.main:main"
